--- apps/app/src/pages/projects.tsx
@@ -1,5 +1,0 @@
-import { FC } from 'react';
-
-export const Projects: FC = () => {
-    return <div className="containerd pt-8">Projects Page</div>;
-};
